requirements.txt
alembic==1.16.1
black==25.1.0
blinker==1.9.0
click==8.2.0
Flask==3.0.3
flask-cors==6.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
